harness_nav/requirements.txt
# Harness Navigation System - Python Dependencies

# GUI Framework
PyQt5>=5.15.0

# Configuration
PyYAML>=6.0

# BeagleBone GPIO (only needed on BeagleBone Black)
# Uncomment when deploying to BeagleBone:
# Adafruit-BBIO>=1.2.0

# Development & Testing
pytest>=7.0.0
pytest-qt>=4.0.0
pytest-cov>=4.0.0

# Type checking
mypy>=1.0.0
types-PyYAML>=6.0.0

# Code formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
